# epita-ing-assistants-yaka-jws-2027-victor.masson/epibazaar/common/src/main/java/fr/epita/assistants/common/api/request/StartRequest.java
package fr.epita.assistants.common.api.request;



import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@AllArgsConstructor
@NoArgsConstructor
@Getter
@Setter
public class StartRequest {
    String mapPath;
}